small fixes

+++ src/cutclasses/ppaccut.cpp
@@ -8,7 +8,7 @@ using namespace std;
 
 void ppaccut::innerloop(treereader *tree,
                              std::vector<std::atomic<bool>> &goodevents,
-                             std::vector<int> range) {
+                             std::vector<uint> range) {
     // Step 1: cloning histograms
     vector<TH1D> _effPPAC;
     vector<vector<TH2D>> _sumdiffppac;
@@ -154,8 +154,8 @@ void ppaccut::analyse(const std::vector<std::string> input, TFile* output){
 
     vector<thread> th;
     for(uint i=0; i<threads; i++){
-        vector<int> ranges = {i*goodevents.size()/threads,
-                              (i+1)*goodevents.size()/threads-1};
+        vector<uint> ranges = {(uint)(i*goodevents.size()/threads),
+                               (uint)((i+1)*goodevents.size()/threads-1)};
         th.emplace_back(thread(&ppaccut::innerloop, this, tree.at(i), ref(goodevents),ranges));
     }
